requirements.txt
numpy==1.24.3
scipy==1.11.1
matplotlib==3.7.2
wfdb==4.1.0
biosppy==1.0.0
pyyaml==6.0.1
scikit-learn==1.3.0
